Refactor option page.
